ix lint warnings around Consensus components
